+++ HorseFluteAnywhere/release-notes.md
@@ -3,7 +3,8 @@
 # Release notes
 ## Upcoming release
 * You can now play the flute by pressing `H` (configurable), without needing to select it on your toolbar.
-* You can now summon the horse without having the horse flute item (if enabled in the `config.json`).
+* You can now summon the horse without having the horse flute item (if enabled).
+* You can now configure the mod in-game if you have [Generic Mod Config Menu](https://www.nexusmods.com/stardewvalley/mods/5098) installed.
 
 ## 1.0
 Released 04 January 2021.